Hehe.. I needed rawContents of a node (who doesnt!)
I'll check to see if this is memory friendly later- but I turned on
the raw function call and switched it to go three levels deep with
xmlNodeDumpOutput, as I couldn't find any htmlNodeDumpFormatOutput.

Also switched out the passing around of the doc pointer, since you
can get to it with the node pointer.

=-Spawrks